Update README Assignment 1.md
After stepping away for a little bit and rereading, I can see how people who are not familiar with this specific software could be confused. There needs to be more detail throughout the entire piece. More purpose of each step and knowing why it is done would have been helpful for readers. While it makes sense in my head, not everyone can decipher everything that I wrote down.  Moving forward with this paper, I would go in and add more details. Pictures would also be very helpful for this to give a more visual component.